pyproject.toml
[project]
name = "kpa-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Pankaj vishw",email = "vishw.dev.1000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-split-settings (>=1.3.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-filter (>=25.1,<26.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "^9.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
